* hypernext:
  ui: user menu: use initals instead of a generic icon (#5351)
  bug/medium: eln import: status and categories for templates
  bug/minor + feat: tinymce code revamp (#5353)
  bug/minor: eln: make .eln valid ro-crate
  dev: fix tests
  bug/medium: eln import: prevent duplication of resources categories
  bug/minor: listing: improve behavior with always show owned parameter
  apidoc: add archived query param for uploads readall
  feat: uploads api: allow filtering archived uploads with api
  dev: no need for the cast to int
  bug/medium + feat: cli timestamp: fix and improve sql query
  feat: import eln: display checksums of failed imported files
  feat: import eln: display error instead of aborting if file shasum fails
  chore: bump elabftw version
  feat: eln export: allow restricting users or resources categories
  chore: composer upgrade
  bug/minor: editor: fix tinymce image plugin missing on mobile-tablets (#5339)
  dev: cypress: increase defaultCommandTimeout from 8s to 15s
  bugfix: minor: fix orcid being unremovable (#5335)

+++ src/Import/Handler.php
@@ -58,7 +58,8 @@ class Handler implements RestInterface
         $Importer->import();
         $inserted = $Importer->getInserted();
         if ($inserted > self::AUDIT_THRESHOLD) {
-            AuditLogs::create(new AuditEventImport($this->requester->userid ?? 0, $inserted));
+            /** @psalm-suppress RedundantCast had an error during eln import where userid was a string for some reason... */
+            AuditLogs::create(new AuditEventImport((int) ($this->requester->userid ?? 0), $inserted));
         }
         return $inserted;
     }
